+++ tests/echidna/properties/DebtTokenProp.sol
@@ -10,7 +10,7 @@ contract DebtTokenProp is PropertiesBase {
 
     /// @custom:invariant 400 - `approveDelegation()` must never revert.
     /// @custom:invariant 401 - Allowance must be modified correctly via `approve()`.
-    function randApproveDelegation(LocalVars_UPTL memory vul, uint seedUser, uint seedSender, uint seedVToken, uint seedAmt) public {
+    function randApproveDelegation(LocalVars_UPTL memory vul, uint8 seedUser, uint8 seedSender, uint8 seedVToken, uint128 seedAmt) public {
         randUpdatePriceAndTryLiquidate(vul);
 
         uint randUser = clampBetween(seedUser, 0 ,totalNbUsers);
